+++ packages/js/src/server/integrations/unhandled_rejection_monitor.ts
@@ -5,10 +5,44 @@ import { Types } from '@honeybadger-io/core'
 export default class UnhandledRejectionMonitor {
   protected __isReporting: boolean
   protected __client: typeof Client
+  protected __listener: (reason: unknown, _promise: Promise<unknown>) => void
 
-  constructor(client: typeof Client) {
+  constructor() {
     this.__isReporting = false
+    this.__listener = this.makeListener()
+  }
+
+  setClient(client: typeof Client) {
     this.__client = client
+  }
+
+  makeListener() {
+    const honeybadgerUnhandledRejectionListener = (reason: unknown, _promise: Promise<unknown>) => {
+      this.__isReporting = true;
+      this.__client.notify(reason as Types.Noticeable, { component: 'unhandledRejection' }, {
+        afterNotify: () => {
+          this.__isReporting = false;
+          if (!this.hasOtherUnhandledRejectionListeners()) {
+            fatallyLogAndExit(reason as Error)
+          }
+        }
+      })
+    }
+    return honeybadgerUnhandledRejectionListener
+  }
+
+  maybeAddListener() {
+    const listeners = process.listeners('unhandledRejection')
+    if (!listeners.includes(this.__listener)) {
+      process.on('unhandledRejection', this.__listener)
+    }
+  }
+
+  maybeRemoveListener() {
+    const listeners = process.listeners('unhandledRejection')
+    if (listeners.includes(this.__listener)) {
+      process.removeListener('unhandledRejection', this.__listener)
+    }
   }
 
   /**
@@ -18,25 +52,8 @@ export default class UnhandledRejectionMonitor {
    * which is to exit the process with code 1
    */
   hasOtherUnhandledRejectionListeners() {
-    return process.listeners('unhandledRejection').length > 1
-  }
-
-  handleUnhandledRejection(reason: unknown, _promise: Promise<unknown>) {
-    if (!this.__client.config.enableUnhandledRejection) {
-      if (!this.hasOtherUnhandledRejectionListeners() && !this.__isReporting) {
-        fatallyLogAndExit(reason as Error)
-      }
-      return
-    }
-
-    this.__isReporting = true;
-    this.__client.notify(reason as Types.Noticeable, { component: 'unhandledRejection' }, {
-      afterNotify: () => {
-        this.__isReporting = false;
-        if (!this.hasOtherUnhandledRejectionListeners()) {
-          fatallyLogAndExit(reason as Error)
-        }
-      }
-    })
+    const otherListeners = process.listeners('unhandledRejection')
+      .filter(listener => listener !== this.__listener)
+    return otherListeners.length > 0
   }
 }