+++ packages/js/src/server/integrations/unhandled_rejection_plugin.ts
@@ -2,16 +2,18 @@ import { Types } from '@honeybadger-io/core'
 import Client from '../../server'
 import UnhandledRejectionMonitor from './unhandled_rejection_monitor'
 
+const unhandledRejectionMonitor = new UnhandledRejectionMonitor()
+
 export default function (): Types.Plugin {
   return {
     load: (client: typeof Client) => {
-      if (!client.config.enableUnhandledRejection) {
-        return
-      }
-      const unhandledRejectionMonitor = new UnhandledRejectionMonitor(client)
-      process.on('unhandledRejection', function honeybadgerUnhandledRejectionListener(reason, _promise) {
-        unhandledRejectionMonitor.handleUnhandledRejection(reason, _promise)
-      })
-    }
+      unhandledRejectionMonitor.setClient(client)
+      if (client.config.enableUnhandledRejection) {
+        unhandledRejectionMonitor.maybeAddListener()
+      } else {
+        unhandledRejectionMonitor.maybeRemoveListener()
+      } 
+    }, 
+    shouldReloadOnConfigure: true,
   }
 }