+++ packages/js/test/e2e/server.js
@@ -1,0 +1,49 @@
+const fs = require('fs')
+const http = require('http')
+const path = require('path')
+
+const PORT = process.env.PORT || 3000;
+
+const MIME_TYPES = {
+  default: 'application/octet-stream',
+  html: 'text/html; charset=UTF-8',
+  js: 'application/javascript',
+  css: 'text/css',
+  png: 'image/png',
+  jpg: 'image/jpg',
+  gif: 'image/gif',
+  ico: 'image/x-icon',
+  svg: 'image/svg+xml',
+};
+
+const STATIC_PATH = path.resolve(__dirname, '../../')
+
+const toBool = [() => true, () => false];
+
+const prepareFile = async (url) => {
+  const paths = [STATIC_PATH, url];
+  if (url === '/') paths.push('test/e2e/sandbox.html')
+
+  const filePath = path.join(...paths)
+  console.log('original url', url, '=> looking for', filePath)
+  const pathTraversal = !filePath.startsWith(STATIC_PATH);
+  const exists = await fs.promises.access(filePath).then(...toBool);
+  const found = !pathTraversal && exists;
+  const streamPath = found ? filePath : STATIC_PATH + '/test/e2e/404.html';
+  const ext = path.extname(streamPath).substring(1).toLowerCase();
+  const stream = fs.createReadStream(streamPath);
+  return { found, ext, stream };
+};
+
+http
+  .createServer(async (req, res) => {
+    const file = await prepareFile(req.url)
+    const statusCode = file.found ? 200 : 404;
+    const mimeType = MIME_TYPES[file.ext] || MIME_TYPES.default;
+    res.writeHead(statusCode, { 'Content-Type': mimeType });
+    file.stream.pipe(res);
+    console.log(`${req.method} ${req.url} ${statusCode}`);
+  })
+  .listen(PORT);
+
+console.log(`Server running at http://127.0.0.1:${PORT}/`);