+++ packages/nextjs/rollup.config.js
@@ -37,8 +37,6 @@ export default {
     commonjs(),
     copy({
       targets: [
-        { src: 'build/copy-config-files.js*', dest: 'dist' },
-        { src: 'build/copy-config-files-exec.js*', dest: 'dist' },
         { src: 'build/*.d.ts', dest: 'dist' },
       ]
     })