+++ packages/gatsby-plugin/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.0.11](https://github.com/honeybadger-io/honeybadger-js/compare/@honeybadger-io/gatsby-plugin-honeybadger@6.0.10...@honeybadger-io/gatsby-plugin-honeybadger@6.0.11) (2023-09-12)
+
+**Note:** Version bump only for package @honeybadger-io/gatsby-plugin-honeybadger
+
+
+
+
+
+## [6.0.10](https://github.com/honeybadger-io/honeybadger-js/compare/@honeybadger-io/gatsby-plugin-honeybadger@6.0.9...@honeybadger-io/gatsby-plugin-honeybadger@6.0.10) (2023-09-07)
+
+**Note:** Version bump only for package @honeybadger-io/gatsby-plugin-honeybadger
+
+
+
+
+
+## [6.0.9](https://github.com/honeybadger-io/honeybadger-js/compare/@honeybadger-io/gatsby-plugin-honeybadger@6.0.8...@honeybadger-io/gatsby-plugin-honeybadger@6.0.9) (2023-09-04)
+
+**Note:** Version bump only for package @honeybadger-io/gatsby-plugin-honeybadger
+
+
+
+
+
 ## [6.0.8](https://github.com/honeybadger-io/honeybadger-js/compare/@honeybadger-io/gatsby-plugin-honeybadger@6.0.7...@honeybadger-io/gatsby-plugin-honeybadger@6.0.8) (2023-08-21)
 
 **Note:** Version bump only for package @honeybadger-io/gatsby-plugin-honeybadger